+++ src/lifecycle/run.rs
@@ -140,16 +140,26 @@ fn run_impl<T: State>(title: &str, size: Vector, settings: Settings) -> Result<(
 
     let key_names = generate_key_names();
     handle_event(&document, &app, move |mut app, event: KeyDownEvent| {
-        let state = ButtonState::Pressed;
+        // If the key press is 'Backspace', we shouldn't send a typed event
+        // However, if it is a single alphanumeric character, that should be a typed event
+        // TODO: this is imperfect at best, a better way to decide what codes get sent to the
+        // application would be desirable
+        let string = event.key();
+        let mut characters = string.chars();
+        let first = characters.next();
+        let second = characters.next();
+        match (first, second) {
+            (Some(ch), None) if ch.is_alphanumeric() => app.event_buffer.push(Event::Typed(ch)),
+            _ => ()
+        }
         if let Some(keycode) = key_names.get(&event.code()) {
-            app.event_buffer.push(Event::Key(KEY_LIST[*keycode], state));
+            app.event_buffer.push(Event::Key(KEY_LIST[*keycode], ButtonState::Pressed));
         }
     });
     let key_names = generate_key_names();
     handle_event(&document, &app, move |mut app, event: KeyUpEvent| {
-        let state = ButtonState::Released;
         if let Some(keycode) = key_names.get(&event.code()) {
-            app.event_buffer.push(Event::Key(KEY_LIST[*keycode], state));
+            app.event_buffer.push(Event::Key(KEY_LIST[*keycode], ButtonState::Released));
         }
     });